+++ TicketPlatFormServer/Repository/Balance/IBalanceRepository.cs
@@ -2,21 +2,21 @@ namespace TicketPlatFormServer.Repository.Balance;
 
 public interface IBalanceRepository
 {
-    Task<DBModel.UserBalance?> GetByUserIdAsync(long userId);
+    Task<DBModel.UserBalance?> GetByUserIdAsync(int userId);
 
-    Task<DBModel.UserBalance> GetOrCreateByUserIdAsync(long userId);
+    Task<DBModel.UserBalance> GetOrCreateByUserIdAsync(int userId);
 
-    Task<int> AtomicDebitAsync(long userId, long amount);
+    Task<int> AtomicDebitAsync(int userId, long amount);
 
-    Task<int> AtomicCreditAsync(long userId, long amount);
+    Task<int> AtomicCreditAsync(int userId, long amount);
 
-    Task<int> AtomicPendingCreditAsync(long userId, long amount);
+    Task<int> AtomicPendingCreditAsync(int userId, long amount);
 
-    Task<int> AtomicPendingDebitAsync(long userId, long amount);
+    Task<int> AtomicPendingDebitAsync(int userId, long amount);
 
-    Task<List<DBModel.BalanceTransaction>> GetByUserIdAsync(long userId, int page, int pageSize);
+    Task<List<DBModel.BalanceTransaction>> GetByUserIdAsync(int userId, int page, int pageSize);
 
-    Task<int> GetTransactionCountByUserIdAsync(long userId);
+    Task<int> GetTransactionCountByUserIdAsync(int userId);
 
     Task AddTransactionAsync(DBModel.BalanceTransaction transaction);
